+++ src/spl.java
@@ -1,21 +1,21 @@
 public class spl {
 
-    public static float[][] inverse_Spl(float[][] matrix){
+    public static double[][] inverse_Spl(double[][] matrix){
         int col = matrix[0].length;
 
-        float[][] inv = invers_identity.invert(operator.delColAt(matrix,col-1));
-        float[][] b_matrix = operator.takeCol(matrix, col-1);
+        double[][] inv = invers_identity.invert(operator.delColAt(matrix,col-1));
+        double[][] b_matrix = operator.takeCol(matrix, col-1);
         operator.displayMatrix(inv);
 
-        float[][] result_Equation_Matrix = operator.multiplyMatrix(inv, b_matrix);
+        double[][] result_Equation_Matrix = operator.multiplyMatrix(inv, b_matrix);
         return result_Equation_Matrix;
 
     }
     
-    public static float[][] gauss_Spl(float[][] m){
+    public static double[][] gauss_Spl(double[][] m){
         int row = m.length;
         int col = row+1;
-        float[][] result = new float[row][1];
+        double[][] result = new double[row][1];
         m = operator.echelonRow(m);
         int subs;
         int distRow;
@@ -31,9 +31,9 @@ public class spl {
     }
 
     // public static void main(String[] args){
-    //     float[][] m = {{1,1,1,6,1},
+    //     double[][] m = {{1,1,1,6,1},
     //     {1,2,3,14,1},{2,3,2,14,1}};
-    //     float[][] res = gauss_Spl(m);
+    //     double[][] res = gauss_Spl(m);
     //     operator.displayMatrix(res);
